Add game loop, and winner screen
Added a game loop, currently set to 1 to test the winner screen. A fade-
out transition was added to remove elements from the screen in a way
that is not so abrubt.

Thee different functions added that check the score, assign a winner and
assign a new element id.

Added a background colour and border to players selection.